refactor: extract SingleCard container
